// src/types.ts
export enum BrushMode {
    HOVER = 'HOVER',
    DRAW = 'DRAW',
    ERASER = 'ERASER'
}

export enum Direction {
    TOP = 'TOP',
    BOTTOM = 'BOTTOM',
    RIGHT = 'RIGHT',
    LEFT = 'LEFT',
    BOTTOM_RIGHT = 'BOTTOM_RIGHT',
    BOTTOM_LEFT = 'BOTTOM_LEFT',
    TOP_RIGHT = 'TOP_RIGHT',
    TOP_LEFT = 'TOP_LEFT'
}